track mplayer config, modify dumbInstall script to ignore symbolic links

+++ dumbInstall.sh
@@ -7,7 +7,7 @@ IGNOREDIR="$DOTDIR/ignore"
 SCRIPTFILE="createSymlinks.sh"
 
 # list of files and directories to process
-LIST="bashrc aliases profile dircolors inputrc xsessionrc Xresources Xresources.d gitconfig gitignore latexmkrc i3status.conf vimrc vim"
+LIST="bashrc aliases profile dircolors inputrc xsessionrc Xresources Xresources.d gitconfig gitignore latexmkrc i3status.conf vimrc vim mplayer"
 CONFIG="i3 user-dirs.dirs user-dirs.locale htop"
 #LIST="file1 file2 dir1 dir2"
 #CONFIG="file3 file4 dir3 dir4"
@@ -24,16 +24,21 @@ for el in $LIST; do
     # name of current element
     #name=$(basename "$f")
 
-    echo "    $el"
+    # only process the current element if it is not a symbolic link
+    if [ ! -L "$HOME/.$el" ]; then
+        echo "    $el"
 
-    # backup the current element if it exists in the user's home directory
-    if [ -f "$HOME/.$el" ] || [ -d "$HOME/.$el" ]; then
-        cp -Lr $HOME/.$el $BACKDIR/$el
-        rm -rf $HOME/.$el
+        # backup the current element if it exists in the user's home directory
+        if [ -f "$HOME/.$el" ] || [ -d "$HOME/.$el" ]; then
+            cp -Lr $HOME/.$el $BACKDIR/$el
+            rm -rf $HOME/.$el
+        fi
+
+        # symlink the current element into the user's home directory
+        ln -s $DOTDIR/$el $HOME/.$el
+    else
+        echo "    $el: symbolic link, do nothing"
     fi
-
-    # symlink the current element into the user's home directory
-    ln -s $DOTDIR/$el $HOME/.$el
 
     # clear temporary variables
     #unset name
@@ -45,16 +50,21 @@ for el in $CONFIG; do
     # name of current element
     #name=$(basename "$f")
 
-    echo "    config/$el"
+    # only process the current element if it is not a symbolic link
+    if [ ! -L "$HOME/.config/$el" ]; then
+        echo "    config/$el"
 
-    # backup the current element if it exists in the user's config directory
-    if [ -f "$HOME/.config/$el" ] || [ -d "$HOME/.config/$el" ]; then
-        cp -Lr $HOME/.config/$el $BACKDIR/config/$el
-        rm -rf $HOME/.config/$el
+        # backup the current element if it exists in the user's config directory
+        if [ -f "$HOME/.config/$el" ] || [ -d "$HOME/.config/$el" ]; then
+            cp -Lr $HOME/.config/$el $BACKDIR/config/$el
+            rm -rf $HOME/.config/$el
+        fi
+
+        # symlink the current element into the user's config directory
+        ln -s $DOTDIR/config/$el $HOME/.config/$el
+    else
+        echo "    config/$el: symbolic link, do nothing"
     fi
-
-    # symlink the current element into the user's config directory
-    ln -s $DOTDIR/config/$el $HOME/.config/$el
 
     # clear temporary variables
     #unset name
